[Silver III] Title: 주유소, Time: 336 ms, Memory: 39212 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -9,8 +9,8 @@ public class Main {
         int N = Integer.parseInt(br.readLine());
         
         //도로의 길이, 리터당 가격 입력받기
-        int[] distances = new int[N - 1];
-        int[] prices = new int[N];
+        long[] distances = new long[N - 1];
+        long[] prices = new long[N];
         String[] roadInput = br.readLine().split(" ");
         String[] cityInput = br.readLine().split(" ");
         for(int i = 0; i < N - 1; i++){
@@ -19,9 +19,9 @@ public class Main {
         for(int i = 0; i < N; i++){
             prices[i] = Integer.parseInt(cityInput[i]);
         }
-        
-        int minCost = 0; //최소 비용
-        int minPrice = prices[0]; //처음 주유소 가격
+
+        long minCost = 0; //최소 비용
+        long minPrice = prices[0]; //처음 주유소 가격
 
         for(int i = 0; i < N - 1; i++){
             if(prices[i] < minPrice) {
